app/src/requirements.txt
Flask==2.3.5
psycopg2-binary==2.9.9

